[Refactor][Kernel] Drop _fwd suffix from mamba kernel filenames
Rename da_cumsum_fwd.py, ssd_chunk_scan_fwd.py, ssd_chunk_state_fwd.py,
ssd_state_passing_fwd.py to match ops-side naming convention.

Part 3 of #924.
